requirements.txt
# Telegram Bot SDK
aiogram==3.4.1
aiohttp==3.9.3
aiosqlite==0.19.0

# Web server for webhook
Flask==3.0.2
gunicorn==21.2.0

# HTTP requests (for sync operations)
requests==2.31.0

# Database
sqlite3 (built-in)

# Utilities
python-dotenv==1.0.0
APScheduler==3.10.4

# Logging and monitoring
python-json-logger==2.0.7

# Type hints
typing-extensions==4.9.0

# For crypto operations
cryptography==42.0.0

# For data validation
pydantic==2.6.0
